adds style for language btn
1. переименована кнопка регистрации
2. флаг вынесен в отдельный компонент
3. добавлены  мои задачи в модалку команды
