+++ src/components/ChangeLanguageBtn/ChangeLanguageBtn.styled.jsx
@@ -1,0 +1,12 @@
+
+import styled from 'styled-components';
+
+export const LanguageBox = styled.button`
+width: 24px;
+height: 100%;
+padding-top: 5px;
+border-radius: 0;
+margin-right: 8px;
+cursor: pointer;
+background: none;
+`;
