+++ src/components/HomePageHeader/HomePageHeader.styled.jsx
@@ -119,9 +119,8 @@ export const Login = styled(NavLink)`
 export const LanguageContainer = styled.div`
 position: absolute;
 top: 30px;
-    right: 30px;
+right: 30px;
 width: 24px;
-height: 100%;
 border-radius: 0;
 margin-right: 8px;
 cursor: pointer;
