+++ src/constants/locales/uk/translation.json
@@ -44,7 +44,7 @@
     "btn": {
         "LogOut": "Вийти",
         "LogIn": "Увійти",
-        "SignUp": "Зареєструватися",
+        "SignUp": "Реєстрація",
         "AddTask": "Додати задачу"
     },
     "calendar": {
